requirements.txt
opencv-python
numpy
mss
pystray
Pillow